+++ example/lib/main.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:weight_scale/scale.dart';
+import 'package:weight_scale/weight_scale.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();