+++ lib/src/features.dart
@@ -1,4 +1,4 @@
-import 'package:weight_scale/scale.dart';
+import 'package:weight_scale/weight_scale.dart';
 
 abstract class BatteryLevelFeature {
   /// The battery level from 0 to 100.