+++ lib/src/weight_scale.dart
@@ -1,4 +1,4 @@
-import 'package:weight_scale/scale.dart';
+import 'package:weight_scale/weight_scale.dart';
 
 /// A connectable weight scale.
 ///