+++ test/src/recognizers/climbro_recognizer_test.dart
@@ -1,5 +1,5 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:weight_scale/scale.dart';
+import 'package:weight_scale/weight_scale.dart';
 import 'package:weight_scale/src/recognizers/climbro_recognizer.dart';
 
 import '../fake_ble_device.dart';