+++ test/src/recognizers/mi_scale_2_recognizer_test.dart
@@ -1,7 +1,7 @@
 import 'dart:typed_data';
 
 import 'package:flutter_test/flutter_test.dart';
-import 'package:weight_scale/scale.dart';
+import 'package:weight_scale/weight_scale.dart';
 import 'package:weight_scale/src/recognizers/mi_scale_2_recognizer.dart';
 
 import '../fake_ble_device.dart';