+++ test/src/scales/mi_scale_2_test.dart
@@ -3,7 +3,7 @@ import 'dart:typed_data';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/annotations.dart';
 import 'package:mockito/mockito.dart';
-import 'package:weight_scale/scale.dart';
+import 'package:weight_scale/weight_scale.dart';
 import 'package:weight_scale/src/scales/mi_scale_2.dart';
 
 @GenerateNiceMocks([MockSpec<BleDevice>()])